requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
Deprecated==1.2.13
gevent==21.12.0
greenlet==1.1.2
idna==3.3
numpy==1.21.5
packaging==21.3
pyparsing==3.0.6
PyYAML==6.0
redis==4.1.0
requests==2.26.0
urllib3==1.26.7
wrapt==1.13.3
zope.event==4.5.0
zope.interface==5.4.0
